backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyjwt==2.9.0
bcrypt==4.2.0
pydantic==2.9.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
websockets==12.0